backend/api/rest.py
import logging
import traceback

from flask_restplus import Api
from backend import settings
from sqlalchemy.orm.exc import NoResultFound

log = logging.getLogger(__name__)

api = Api(version='1.0', title='Chatbot Analytics Api',
          description='A simple demonstration of a chatbot analysis on machine learning platform')


@api.errorhandler
def default_error_handler(e):
    message = 'An unhandled exception occurred.'
    log.exception(message)

    if not settings.FLASK_DEBUG:
        return {'message': message}, 500


@api.errorhandler(NoResultFound)
def database_not_found_error_handler(e):
    log.warning(traceback.format_exc())
    return {'message': 'A data result was required but none was found.'}, 404
