requirements.txt
Flask==1.0.2
Flask-Cors==3.0.7
numpy
pytest==4.2.0
pandas==0.24.1
flask_marshmallow==0.9.0
flask-restplus==0.12.1
Flask-SQLAlchemy==2.3.2
Werkzeug==0.16.1
gunicorn
Flask-PyMongo
openpyxl
requests